+++ driver_mobile/lib/widgets/custom_button.dart
@@ -63,7 +63,9 @@ class CustomButton extends StatelessWidget {
         button = ElevatedButton(
           onPressed: activeOnPressed,
           style: height != null
-              ? ElevatedButton.styleFrom(minimumSize: Size(width ?? double.infinity, height!))
+              ? ElevatedButton.styleFrom(
+                  minimumSize: Size(width ?? double.infinity, height!),
+                )
               : null,
           child: label,
         );
@@ -72,7 +74,9 @@ class CustomButton extends StatelessWidget {
         button = OutlinedButton(
           onPressed: activeOnPressed,
           style: height != null
-              ? OutlinedButton.styleFrom(minimumSize: Size(width ?? double.infinity, height!))
+              ? OutlinedButton.styleFrom(
+                  minimumSize: Size(width ?? double.infinity, height!),
+                )
               : null,
           child: label,
         );
@@ -81,7 +85,9 @@ class CustomButton extends StatelessWidget {
         button = TextButton(
           onPressed: activeOnPressed,
           style: height != null
-              ? TextButton.styleFrom(minimumSize: Size(width ?? double.infinity, height!))
+              ? TextButton.styleFrom(
+                  minimumSize: Size(width ?? double.infinity, height!),
+                )
               : null,
           child: label,
         );
@@ -89,10 +95,7 @@ class CustomButton extends StatelessWidget {
     }
 
     if (width != null) {
-      return SizedBox(
-        width: width,
-        child: button,
-      );
+      return SizedBox(width: width, child: button);
     }
     return button;
   }